# compaction: merge in-use key ranges level-by-level