When computing in-use key ranges for a compaction, merge the intervals
for each level before descending to the next level. This allows for an
optimization where a level iterator may be seeked to the current
accumulated key range's end boundary. Since higher-level tables tend to
be broader, this should help reduce per-sstable work when setting up
flushes/compactions in stores with many sstables.

Pebble benchmarks are largely even.

```
name                old ops/sec  new ops/sec  delta
ycsb/D/values=1024    209k ±10%    206k ± 6%    ~     (p=0.312 n=10+14)
ycsb/C/values=1024    895k ±13%    940k ±16%    ~     (p=0.103 n=10+15)
ycsb/A/values=1024   80.8k ± 1%   80.6k ± 2%    ~     (p=0.643 n=10+15)
ycsb/B/values=1024    383k ± 7%    384k ± 7%    ~     (p=0.978 n=10+15)
ycsb/E/values=1024   63.1k ± 9%   62.8k ± 7%    ~     (p=0.935 n=10+15)
ycsb/F/values=1024   26.9k ± 3%   27.0k ± 2%    ~     (p=0.605 n=10+15)
ycsb/F/values=64      239k ± 1%    238k ± 0%  -0.46%  (p=0.021 n=10+12)
ycsb/C/values=64     2.32M ±24%   2.55M ±37%    ~     (p=0.177 n=10+15)
ycsb/E/values=64      212k ± 6%    215k ± 5%    ~     (p=0.261 n=10+15)
ycsb/B/values=64     1.11M ± 8%   1.13M ±16%    ~     (p=0.428 n=10+15)
ycsb/A/values=64      824k ± 6%    846k ± 6%    ~     (p=0.235 n=10+14)
ycsb/D/values=64      914k ± 4%    935k ±14%    ~     (p=0.194 n=9+15)
```

It's difficult to write a natural microbenchmark for this because
flushes or compactions will change the shape of the LSM during the benchmark,
and it's costly to set up the requisite number of files. I added a
really contrived benchmark that demonstrates the extreme best case for
the optimization, where one files span across all the other files.

```
name                                                      old time/op    new time/op    delta
ManySSTables/calculateInuseKeyRanges/sstables=10-24         4.48µs ± 1%    3.22µs ± 1%   -28.15%  (p=0.008 n=5+5)
ManySSTables/calculateInuseKeyRanges/sstables=1000-24        272µs ± 1%       4µs ± 1%   -98.61%  (p=0.008 n=5+5)
ManySSTables/calculateInuseKeyRanges/sstables=10000-24      3.74ms ± 1%    0.00ms ± 1%   -99.90%  (p=0.008 n=5+5)
ManySSTables/calculateInuseKeyRanges/sstables=100000-24     53.3ms ± 2%     0.0ms ± 1%   -99.99%  (p=0.008 n=5+5)
ManySSTables/calculateInuseKeyRanges/sstables=1000000-24     826ms ±34%       0ms ± 5%  -100.00%  (p=0.008 n=5+5)

name                                                      old alloc/op   new alloc/op   delta
ManySSTables/calculateInuseKeyRanges/sstables=10-24         3.15kB ± 0%    1.87kB ± 0%   -40.61%  (p=0.008 n=5+5)
ManySSTables/calculateInuseKeyRanges/sstables=1000-24        100kB ± 0%       2kB ± 0%   -98.13%  (p=0.008 n=5+5)
ManySSTables/calculateInuseKeyRanges/sstables=10000-24      2.57MB ± 0%    0.00MB ± 0%   -99.93%  (p=0.000 n=4+5)
ManySSTables/calculateInuseKeyRanges/sstables=100000-24     26.6MB ± 0%     0.0MB ± 0%   -99.99%  (p=0.008 n=5+5)
ManySSTables/calculateInuseKeyRanges/sstables=1000000-24     252MB ± 0%       0MB ± 0%  -100.00%  (p=0.008 n=5+5)

name                                                      old allocs/op  new allocs/op  delta
ManySSTables/calculateInuseKeyRanges/sstables=10-24           20.0 ± 0%      19.0 ± 0%    -5.00%  (p=0.008 n=5+5)
ManySSTables/calculateInuseKeyRanges/sstables=1000-24         26.0 ± 0%      19.0 ± 0%   -26.92%  (p=0.008 n=5+5)
ManySSTables/calculateInuseKeyRanges/sstables=10000-24        36.0 ± 0%      19.0 ± 0%   -47.22%  (p=0.008 n=5+5)
ManySSTables/calculateInuseKeyRanges/sstables=100000-24       47.4 ± 1%      14.0 ± 0%   -70.46%  (p=0.008 n=5+5)
ManySSTables/calculateInuseKeyRanges/sstables=1000000-24      76.4 ± 9%      19.0 ± 0%   -75.13%  (p=0.008 n=5+5)
```

Fix #718.